condense files and allow for multiple tokens
